feat: Add  get commands
This commit adds the version and get commands to the kubectl-sg4pod CLI tool. The version command prints the version number of kubectl-sg4pod, while the get command retrieves resources such as Pods from the Kubernetes cluster. The get command fetches a list of Pods from the specified namespace or the default namespace if none is specified. It also displays the status and IP addresses of the Pods.