feat: Add RunGetPods function to retrieve and print pod information
This commit adds a new function called RunGetPods to the k8s package. This function retrieves the pod information from the specified namespace and prints it in a table format using tabwriter. It uses the Kubernetes clientset to interact with the cluster and fetches the list of pods using the List method. The pod name, IP address, and status are printed in the table.

The purpose of this change is to provide a convenient way to retrieve and display pod information in a readable format.

Refs: #7