add-output-method (#9)
* feat: Add RunGetPods function to retrieve and print pod information

This commit adds a new function called RunGetPods to the k8s package. This function retrieves the pod information from the specified namespace and prints it in a table format using tabwriter. It uses the Kubernetes clientset to interact with the cluster and fetches the list of pods using the List method. The pod name, IP address, and status are printed in the table.

The purpose of this change is to provide a convenient way to retrieve and display pod information in a readable format.

Refs: #7

* feat: Add RunGetPods function and update package name and binary name

This commit adds the RunGetPods function to retrieve and print pod information. It also updates the package name and binary name. These changes improve the functionality and organization of the codebase.

* feat: Add install target to Makefile for easy installation

Add an install target to the Makefile to make it easier to install the binary. The install target builds the binary and moves it to the /usr/local/bin directory using the sudo command.