Refactor by kubectl plugin builder (#13)
* refactor: rename plugin

* refactor: rename plugin

* refactor: apply kubectl plugin builder

* docs: add README

* fix: add test option

* chore: bumpup go version

* refactor: apply kubectl plugin builder

* refactor: apply kubectl plugin builder to go.mod

This commit applies the kubectl plugin builder to the go.mod file. It updates the k8s.io/client-go dependency to version 0.31.1.

- Refactor go.mod to apply kubectl plugin builder
- Update k8s.io/client-go to version 0.31.1

* refactor: add unit test for OutputYAML function

* refactor: update OutputJSON function to accept generic data

This commit updates the `OutputJSON` function in the `json.go` file to accept generic data instead of specifically requiring a slice of `PodInfo`. It also adds input validation for the data. The function now marshals the data into JSON format and writes it to the output writer.

Resolves: #issue_number